refactor: HelmPane and BoardKeys
