+++ src/components/organisms/KeyboardShortcuts/BoardKeys/BoardKeys.styled.tsx
@@ -16,7 +16,7 @@ export const StyledShortCell = styled.div`
 `;
 
 export const StyledKey = styled.div`
-  display: flex;
+  display: flex
   flex-direction: column;
   justify-content: center;
   align-items: center;
